poc/requirements.txt
deepface==0.0.93
numpy==1.24.3
pandas==2.1.2
opencv-python==4.8.1.78
tensorflow==2.13.0
keras==2.13.1
gdown==4.7.1
Pillow==11.0.0
scipy==1.11.2
